prefer static_map! over explicit unsafe in serde deserialisation
Instead of going over all the entries and checking
if they are initialised and then going over them
again and unwrapping unsafely, we can just go over
it once and report the error on first failure. The
semantics should be exactly the same but we no
longer have an explicit `unsafe` in here: it's
moved behind the `static_map!` macro instead which
can be considered more battle-hardened.

If we're accepting unsafe then it might make more
sense to do something like `[MaybeUninit<T>;
L::LENGTH]` though something like the `Builder` is
needed to make this ergonomic. Either way, I think
the version in this commit at least hides the
magic.

PS: I checked if we can just replace
`.unwrap_unchecked()` with `.unwrap()` but it
seems that rustc is not able to notice that it
will never fail so it generates panic code anyway.